Rename azure_monitor_activity_log_alert(s) resources to azurerm_monitor_activity_log_alert(s)
* Add deprecation warnings to azure_monitor_activity_log_alert(s)
* Update docs to reference new resource names
* Update integration tests to use renamed resources
