+++ libraries/azure_monitor_activity_log_alert.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require 'azurerm_resource'
+require 'azurerm_monitor_activity_log_alert'
 
-class AzureMonitorActivityLogAlert < AzurermResource
+class AzureMonitorActivityLogAlert < AzurermMonitorActivityLogAlert
   name 'azure_monitor_activity_log_alert'
-  desc 'Verifies settings for a Azure Monitor Activity Log Alert'
+  desc '[DEPRECATED] Please use the azurerm_monitor_activity_log_alert resource'
   example <<-EXAMPLE
     describe azure_monitor_activity_log_alert(resource_group: 'example', name: 'AlertName') do
       it { should exist }
@@ -12,44 +12,8 @@ class AzureMonitorActivityLogAlert < AzurermResource
     end
   EXAMPLE
 
-  ATTRS = {
-    name:       'name',
-    id:         'id',
-    conditions: 'conditionAllOf',
-    operations: 'operations',
-  }.freeze
-
-  attr_reader(*ATTRS.keys)
-
   def initialize(resource_group: nil, name: nil)
-    resp = client.activity_log_alert(resource_group, name)
-    return if resp.nil? || resp.key?('error')
-
-    @name       = resp['name']
-    @id         = resp['id']
-    @conditions = resp['properties']['condition']['allOf']
-    @operations = collect_operations(@conditions)
-
-    ATTRS.each do |attr_name, api_name|
-      next if instance_variable_defined?("@#{attr_name}")
-
-      instance_variable_set("@#{name}", fields[api_name])
-    end
-
-    @exists = true
-  end
-
-  def to_s
-    "#{name} Activity Log Alert"
-  end
-
-  private
-
-  # Collect all Operation strings for the Activity Log Alert
-  #
-  # @param [Hash] 'allOf' conditions from response properties
-  # @return [Array] of operation strings
-  def collect_operations(conditions)
-    conditions.find_all { |x| x['field'] == 'operationName' }.collect { |x| x['equals'] }
+    warn '[DEPRECATION] The `azure_monitor_activity_log_alert` resource is deprecated and will be removed in version 2.0. Use the `azurerm_monitor_activity_log_alert` resource instead.'
+    super
   end
 end
