--- test/integration/verify/controls/azure_monitor_activity_log_alerts.rb
@@ -1,5 +1,0 @@
-control 'azure_monitor_activity_log_alerts' do
-  describe azure_monitor_activity_log_alerts do
-    its('names') { should include('defaultLogAlert_5_3') }
-  end
-end
