+++ test/integration/verify/controls/azurerm_monitor_activity_log_alerts.rb
@@ -1,0 +1,5 @@
+control 'azurerm_monitor_activity_log_alerts' do
+  describe azurerm_monitor_activity_log_alerts do
+    its('names') { should include('defaultLogAlert_5_3') }
+  end
+end
